Make Executer properly deal with end of list
When a null value is encountered, signifying the end of the list the
number of robots is decremented and the mapping for that robot's list
removed.

Now when adding the commands to the map the BtServer increments the number
of robots and sends a check command to the robot prompting it to reply
ready to the server.